Extract test utils to own folder
This paves the way for adding more test utils
